🌱 [scanner] fix: split oversized dashboard and mission-control components
Extract types, constants, and utils from FlightPlanBlueprint.tsx into
sibling files (FlightPlanBlueprint.types.ts, FlightPlanBlueprint.constants.ts,
FlightPlanBlueprint.utils.ts) to bring effective line count from 802 to
781 (under the 800-line max-lines limit).

Remove eslint-disable max-lines from BlueprintInfoPanels.tsx — the file
was already at 735 effective lines, well under the limit.

No behaviour change — structural refactor only.

Fixes #21500
