+++ web/src/components/mission-control/FlightPlanBlueprint.constants.ts
@@ -1,0 +1,19 @@
+/** Minimum info-panel width (px) */
+export const INFO_PANEL_MIN = 280
+/** Maximum info-panel width (px) */
+export const INFO_PANEL_MAX = 600
+/** Default info-panel width (px) — 26rem */
+export const INFO_PANEL_DEFAULT = 416
+/** localStorage key for persisted panel width */
+export const INFO_PANEL_LS_KEY = 'mission-control-info-panel-width'
+
+export const ZOOM_MIN = 0.3
+export const ZOOM_MAX = 3
+export const ZOOM_STEP = 0.2
+
+/** Minimum gap (SVG units) between two label slots to avoid overlap */
+export const MIN_LABEL_GAP = 14
+/** Radius (SVG units) of a project node — used to push labels clear of nodes */
+export const NODE_RADIUS = 18
+/** Vertical offset (SVG units) to place the label above the edge midpoint */
+export const LABEL_OFFSET_Y = 12
