+++ web/src/components/mission-control/FlightPlanBlueprint.utils.ts
@@ -1,0 +1,9 @@
+import type { PayloadProject } from './types'
+
+/** Resolve kbPath for a project — tries explicit kbPath, then convention-based lookup */
+export function resolveKbPath(proj: PayloadProject): string | undefined {
+  if (proj.kbPath) return proj.kbPath
+  // Convention: fixes/cncf-install/install-{name}.json
+  const slug = proj.name.toLowerCase().replace(/\s+/g, '-')
+  return `fixes/cncf-install/install-${slug}.json`
+}
